requirements.txt
Django==3.0.6
django-summernote==0.8.11.6
django-taggit==1.3.0
Pillow==7.1.2
python-decouple==3.3
django-bootstrap4==2.0.1